refactor: break up image rendering into different files
